backend/server/requirements.txt
flask>=2.0.2
numpy
openai
pandas
tiktoken
flask-cors
kafka-python
